+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Removed any unix socket code when building on Windows.
 - Lots of internal code cleanup/refactoring/restructuring.
 - Exit client with non-zero Exitcode when getting a failure message from the daemon.
+- Make the response of `group` without any arguments more beautiful
 
 ### Added